+++ govtool/frontend/src/components/organisms/StatusModal.tsx
@@ -8,11 +8,16 @@ import { useScreenDimension, useTranslation } from "@/hooks";
 
 export interface StatusModalState {
   buttonText?: string;
+  cancelText?: string;
+  feedbackText?: string;
   status: "warning" | "info" | "success";
   isInfo?: boolean;
   link?: string;
+  linkText?: string;
   message: React.ReactNode;
   onSubmit?: () => void;
+  onCancel?: () => void;
+  onFeedback?: () => void;
   title: string;
   dataTestId: string;
 }
@@ -43,20 +48,20 @@ export function StatusModal() {
           textAlign="center"
           sx={{ fontSize: "16px", fontWeight: "400" }}
         >
-          {state?.message}{" "}
+          {state?.message}
           {state?.link && (
             <Link
               onClick={() => openInNewTab(state?.link || "")}
               target="_blank"
               sx={[{ "&:hover": { cursor: "pointer" } }]}
             >
-              {t("thisLink")}
+              {state?.linkText || t("thisLink")}
             </Link>
           )}
         </Typography>
       </ModalContents>
       <Button
-        data-testid={"confirm-modal-button"}
+        data-testid="confirm-modal-button"
         onClick={state?.onSubmit ? state?.onSubmit : closeModal}
         sx={{
           borderRadius: 50,
@@ -64,11 +69,46 @@ export function StatusModal() {
           padding: "10px 26px",
           textTransform: "none",
           marginTop: "38px",
+          width: "100%",
         }}
         variant="contained"
       >
         {state?.buttonText || t("confirm")}
       </Button>
+      {state?.cancelText && (
+        <Button
+          data-testid="cancel-modal-button"
+          onClick={state?.onCancel ? state?.onCancel : closeModal}
+          sx={{
+            borderRadius: 50,
+            margin: "0 auto",
+            padding: "10px 26px",
+            textTransform: "none",
+            marginTop: "24px",
+            width: "100%",
+          }}
+          variant="outlined"
+        >
+          {state?.cancelText}
+        </Button>
+      )}
+      {state?.feedbackText && (
+        <Button
+          data-testid="feedback-button"
+          onClick={state?.onFeedback ? state?.onFeedback : closeModal}
+          sx={{
+            borderRadius: 50,
+            margin: "0 auto",
+            padding: "10px 26px",
+            textTransform: "none",
+            marginTop: "24px",
+            width: "100%",
+          }}
+          variant="outlined"
+        >
+          {state?.feedbackText}
+        </Button>
+      )}
     </ModalWrapper>
   );
 }