+++ govtool/frontend/src/consts/governanceAction/metadataHashValidationErrors.ts
@@ -1,0 +1,6 @@
+export enum MetadataHashValidationErrors {
+  INVALID_URL = "Invalid URL",
+  INVALID_JSON = "Invalid JSON",
+  INVALID_HASH = "Invalid hash",
+  FETCH_ERROR = "Error fetching data",
+}